requirements.txt
numpy
scikit-learn==1.3.2
scipy
seaborn
streamlit
pandas==2.2.2
pillow==10.4.0
plotly==5.22.0
-e .